Move tests to be integration tests
